ml/requirements.txt
joblib==1.5.2
lightgbm==4.6.0
numpy==2.3.3
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
wheel==0.45.1
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
xgboost>=1.7.0
tpot>=0.12.0
imbalanced-learn>=0.11.0